requirements.txt
# ═══════════════════════════════════════════════════════════
# SIEM Personal v4.0 — Dependencias
# Instalar: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════
pywin32>=306
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
python-nmap>=0.7.1
